+++ lib/glimmer/dsl/xml/tag_expression.rb
@@ -1,4 +1,5 @@
 require 'glimmer/dsl/xml/node_parent_expression'
+require 'glimmer/dsl/xml/xml_expression'
 require 'glimmer/dsl/static_expression'
 require 'glimmer/dsl/top_level_expression'
 require 'glimmer/xml/node'
@@ -11,10 +12,13 @@ module Glimmer
         include NodeParentExpression
 
         def can_interpret?(parent, keyword, *args, &block)
+          hash_arg = args.detect {|arg| arg.is_a?(Hash)}
           (parent == nil or parent.is_a?(Glimmer::XML::Node)) and
           (keyword.to_s == "tag") and
-          ((args.size == 1) and ((args[0].is_a?(Hash)) or (args[0].is_a?(Hash)))) and
-          args[0].include?(:_name)
+            ((XmlExpression::SUPPORTED_ARG_TYPES & args.map(&:class).uniq) == args.map(&:class).uniq) and
+            args.map(&:class).count(Hash) <= 1 and
+            ((args.index(hash_arg) == args.size - 1) or args.index(hash_arg).nil?) and
+            args[0].include?(:_name)
         end
 
         def interpret(parent, keyword, *args, &block)